@@ -19,11 +19,12 @@
     "@angular/platform-browser": "^8.0.0",
     "@angular/platform-browser-dynamic": "^8.0.0",
     "@angular/router": "^8.0.0",
+    "bootstrap": "3.3.7",
     "core-js": "^3.1.2",
+    "http-status-codes": "1.3.2",
     "rxjs": "^6.0.0",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26",
-    "bootstrap": "3.3.7"
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.800.0",
@@ -44,6 +45,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "3.4.5"
+    "typescript": "3.5.2"
   }
 }
